Create Simple calculation
Given a simple equation, you should output the correct value for variable "x". The equation has two numbers greater than 0 and variable "x", and between these can be "+", "-" or "=". Numbers, variable "x", symbols "+", "-", "=" all separated by a space.

It is guaranteed that the equation is correct. The result should be an integer.