Create Multiset
A multiset is a generalization of the concept of a set. Unlike sets, it can store duplicate elements. The number of instances of an element is the multiplicity.

For example, given the following multiset: {a, a, b, b, b, c}
The multiplicity of a is 2, the multiplicity of b is 3, the multiplicity of c is 1. If a multiset does not have an element, the multiplicity of it is 0.

Write an implementation of the provided generic interface Multiset. The template for your generic class named HashMultiset is given as well. You should implement all methods of the class, according to its interface. You also can add additional methods for helping.

Read the given interface to understand the common multiset operations (add, remove, union, intersection and so one).