Create Check Sudoku
N-size sudoku is a game with a square table of N2 width and height divided into N2 smaller squares of N width and height. In a solved state, each of this smaller squares, as well as each row and column of a full square, contains all numbers from 1 to N2 without repetition.

Given a number N on the first line and a full sudoku table on the next N2 lines. Every line contains N2 integers.
Your task is to determine whether this sudoku is solved or not. Output "YES" if this sudoku table is solved, otherwise "NO".

N can be from 1 to 10.