Create The pre-order traversal
A full ternary tree with the depth n is given. Write a program that will print all nodes of the tree in Pre-Order. Enumeration starts from 0 and goes sequentially through tree’s levels. On the picture below you can see a tree with depth 2.

Pre-order for k-ary tree is (Root, Child_1, …., Child_k).

Input: n - the depth of the tree.
Output: Sequence of nodes in Pre-Order, divided by spaces.