Create The number of inversions
The first line contains number 1≤n≤10, second one — array A[1…n], containing natural numbers not greater than 10^9. You need to calculate the number of pairs of indices 1≤i<j≤n, for which A[i]>A[j]. 

Such pair of elements is called the inversion. The number of inversion in the array is in some way its measure of random nature: for example, there are no inversions at all in an array arranged in a non-decreasing order, and in an array, arranged in descending order, every two elements form an inversion.

The result should be long, because an array can contain a lot of inversions.
For example, the sequence 2, 4, 1, 3, 5 has three inversions: (2, 1), (4, 1), (4, 3).

Modify merge sort and solve the problem using it.