Create The full binary tree
A tree is represented by pairs of adjacent nodes; node 0 is the root. The task is to determine whether the given tree is a full binary tree or not. A full binary tree is a binary tree where all levels are full.
It is guaranteed that the input defines a correct tree.

Input: n - the number of edges in the tree. Following that are n lines, describe the edges in the tree using ‘parent-children’ structure.
Output: “yes”, if the given tree is full binary, and “no” otherwise.
Restrictions: n is guaranteed to be no more than 100. The numbers of nodes are also no more than 100.

Sample Input 1:
2
0 1
0 2

Sample Output 1:
yes