Create Fishing Championship
The Vikings have organized a fishing championship. They don’t have to fish every day, but each day they can catch just one fish (or none). Every day they write down their name, date and the result of their fishing on a piece of paper and put this sheet into one large box. Each Viking wants to know what was the longest consecutive series of days when he successfully caught some fish (the series is broken on a day when either the Viking doesn't catch any fish or doesn't go fishing at all) as well as his average number of fish caught per day. Unfortunately, these sheets of paper got mixed up in the box.

Write a program to calculate the longest series and the average result for each Viking.

Input
Each line of the standard input contains the name of the Viking, the date and the results of the fishing expedition (“1” if a Viking caught a fish, and “0” otherwise), separated by a single space.
The number of lines can be between 1 and 50 (inclusive).

Output
To the standard output, print the name of the Viking, an integer denoting his longest series of days in which he caught a fish every day and the average number of fish caught per day (separated by a single space), with one Viking per line.
Vikings should be sorted by their longest series of days when they caught a fish, and the average number of fish caught per day, in descending order.

Constraints
The name of the Viking doesn’t contain any spaces.
The date is written in MM/DD/YYYY format (i.e. 02/28/2019).
The average number of fish caught per day should be rounded to a number with two decimal places.

Example input
Bardisson 02/19/2019 0
Bardisson 02/20/2019 1
Nikolasson 02/20/2019 1
Bardisson 02/21/2019 1
Nikolasson 02/21/2019 0
Hrollaug 02/21/2019 0
Example output
Bardisson 2 0.67
Nikolasson 1 0.5
Hrollaug 0 0