Create Ternary operator
Let's combine function to build more complex functions using standard functional interfaces.

Given the predicate condition and two functions ifTrue and ifFalse. Implement a method ternaryOperator which builds a new function returning the result of invoking ifTrue, if the predicate is true and otherwise - the result of invoking ifFalse.

Here is an example:
Predicate<Object> condition = Objects::isNull;
Function<Object, Integer> ifTrue = obj -> 0;
Function<CharSequence, Integer> ifFalse = CharSequence::length;
Function<String, Integer> safeStringLength = ternaryOperator(condition, ifTrue, ifFalse);
In the example above, the resulting function returns the length of a string if the reference to the string is not null, otherwise, it returns 0.

Note, it can be collapsed into a single line, but for clarity, additional variables are used.