Create The depth of the nodes
There is an empty binary search tree. You receive two types of queries: add a node with a given key and determine the depth of the node with a given key. For each ‘depth determination’ query print the calculated depth.

Input: On the first line – the number of queries n. On the next n lines there are queries of two types: «+ x» -- add an element with the key ‘x’ to the binary search tree and «? x» - to determine the depth of the vertex with the 'x' key
Output: For all queries of «? x» type print the depth of the node with the key ‘x’ in the new line. If there is no such node, print «no».

Sample Input 1:
10
+ 2
+ 3
+ 5
+ 1
+ 4
? 2
? 3
? 5
? 1
? 4

Sample Output 1:
0
1
2
1
3