Create The height of the tree
There is an empty binary search tree and one operation on him – node addition. A sequence of keys is given. What will be the height of the tree after adding all the keys in that order?

Input: On the first line – the number of nodes n. On the second line are keys that you need to add to the tree, divided by spaces. It is guaranteed that all the keys are different.
Output: On the first line print one number: the height of the resulting tree.

Sample Input 1:
5
2 3 5 1 4
Sample Output 1:
3