Create Swap binary heaps
Numbers from 1 to n are added to min-heap in a certain order. For each number find out how many times it changed its position in the min-heap. Presume that you use the code that was given in the first part of the lesson.
Clarification: for addition use method Insert(), add nodes in the same order as they are in the input.

Input: In the first line is the number n. In the second line, divided by spaces, are n numbers from 1 to n.
Output: n numbers divided by spaces: i-th number indicates the number of position changes of the number i in the constructed min-heap.

Sample Input 1:
5
5 4 3 2 1
Sample Output 1:
2 3 3 2 2