Create Sort with States
You're given an array of elements. Your task is to sort it in ascending order using binary heap. Print not only the sorted array, but also the initial state of the heap and its states after extracting every element.

Input: In the first line is N - the length of the array. In the next line are elements of the array - N whole numbers, each is no more than 10^9 in absolute value. (1≤N≤100).
Output: In the first line print the heap that was constructed after calling minHeap(), and in the following N-1 lines print heap states after removing next element. In the last (N+1-th) line print the sorted array.

Sample Input 1:
3
3 2 1
Sample Output 1:
1 2 3
2 3
3
1 2 3